Refactored status command, moved new status argument to flag argument, with validation.
